requirements.txt
Django>=5.0
djangorestframework>=3.14
gunicorn>=21.2
mysqlclient>=2.2.0
